+++ license-manager-client-library/docs/release-notes.rst
@@ -22,6 +22,10 @@
 License Manager Client Library Release Notes
 ============================================
 
+Version 0.0.6, 14 May 2019
+-------------------------------
+* Update license-manager-client-library version as security-verification version changes (`ACUMOS-2887 <https://jira.acumos.org/browse/ACUMOS-2887>`_)
+
 Version 0.0.5, 10 May 2019
 -------------------------------
 * Update license-manager-client-library version as security-verification version changes (`ACUMOS-2887 <https://jira.acumos.org/browse/ACUMOS-2887>`_)